requirements.txt
fastapi[standard]
python-jose
passlib
bcrypt
pymongo
pyjwt